Added Framework System & Framework Render. Modified Render System, Window, Message, and main to accomodate this.
Framework System is a system that allows for the System class to work
low end functions without knowing the low endsystems. The first
framework class implemented is the window and render class. This handles
creation of windows and rendering graphics to windows. Render System
modified to take advantage of the framework system. New message type
added to send framework to systems. Main modified to take advantage of
the new framework system.